Move spam logger to models
